"Basic" Code generation for Probing
A probe is simply a named wire; the idea is to
support getting a state/output by name in the
python script. This would facilitate easy
signal acquisition as well as automated testing
to verify that signals behave as expected.

Right now I only generate a function with the
appropriate signal name to take the solution
to the ODE and produce the signal in question.